src/main/java/com/bestmatchedrestaurantsadrianosotos/demo/core/domains/Restaurant/Restaurant.java
package com.bestmatchedrestaurantsadrianosotos.demo.core.domains.Restaurant;

public class Restaurant {
    private String name;
    private int rating;
    private int distance;
    private int averagePrice;
    private String cuisine;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getRating() {
        return rating;
    }

    public void setRating(int rating) {
        this.rating = rating;
    }

    public int getDistance() {
        return distance;
    }

    public void setDistance(int distance) {
        this.distance = distance;
    }

    public int getAveragePrice() {
        return averagePrice;
    }

    public void setAveragePrice(int averagePrice) {
        this.averagePrice = averagePrice;
    }

    public String getCuisine() {
        return cuisine;
    }

    public void setCuisine(String cuisine) {
        this.cuisine = cuisine;
    }
}
